Optimize : don't discriminate "(non)exact" for Deadlock
We have special code that is used to make fixpoint over PredRel more
efficient, but it could cost a lot (proportional to number of events) if
we are only doing one step (e.g. deadlock) it is not worth it to compute
that.
Improves situation for e.g. BART-PT model deadlock detection.